# Repository/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Matrix Repository Api",
    "description": "API to create and receive computed matrices",
    "version": "1.0.0"
  },
  "paths": {
    "/matrices": {
      "post": {
        "tags": [
          "Repository, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "CreateMatrix",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatrixCreationOptions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixCreationResult"
                }
              }
            }
          }
        }
      }
    },
    "/matrices/{id}/computed": {
      "get": {
        "tags": [
          "Repository, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "IsMatrixComputed",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputationState"
                }
              }
            }
          }
        }
      }
    },
    "/matrices/{id}": {
      "get": {
        "tags": [
          "Repository, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "GetMatrix",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ComputationState": {
        "type": "object",
        "properties": {
          "matrixId": {
            "type": "integer",
            "format": "int32"
          },
          "isComputed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "MatrixCreationOptions": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "integer",
            "format": "int32"
          },
          "columns": {
            "type": "integer",
            "format": "int32"
          },
          "hash": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MatrixCreationResult": {
        "type": "object",
        "properties": {
          "matrixId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}